support-chatbot/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyjwt==2.9.0
httpx==0.27.2
python-multipart==0.0.9
pydantic==2.9.2
